Add a tip about `portable-file-dialogs.h` runtime req on Linux (#358)
In `portable-file-dialogs.h`, if none of `zenity`, `matedialog`,
`qarma`, `kdialog` is installed, it simply runs `echo`[1]. I'm not
sure if the intent was to fallback to readline or something, but it
seems to simply return an empty string with no indication of what
went wrong. This causes SpaghettiKart to render an error box:

> No ROM selected.
>
> Exiting...

To me this seems like a bug in `portable-file-dialogs.h`.

[1]: https://github.com/adsr/SpaghettiKart/blob/11793ed4318b35e205704c46263cc076b97d763a/include/portable-file-dialogs.h#L899

Co-authored-by: MegaMech <MegaMech@users.noreply.github.com>